Get articles to appear on refresh (#74)
* Get articles to appear on refresh

* Replace Coil with Glide

+++ basil/src/test/java/com/jocmp/basil/persistence/FeedRecordsTest.kt
@@ -27,7 +27,7 @@ class FeedRecordsTest {
 
         val result = database
             .articlesQueries
-            .findBy(articleID = article.id.toLong())
+            .findBy(articleID = article.id)
             .executeAsOneOrNull()
 
         assertNull(result)
